refactor(RestaurantRequest): deliveryFee 자료형 변경

+++ src/main/java/com/flab/woowahaneats/domain/restaurant/controller/dto/RestaurantRequest.java
@@ -4,8 +4,6 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
-
 public record RestaurantRequest(
 
         @NotNull
@@ -27,7 +25,7 @@ public record RestaurantRequest(
         int minOrderAmt,
 
         @NotNull
-        BigDecimal deliveryFee,
+        int deliveryFee,
 
         @NotBlank
         String region
